pytorch-fashion-classifier/requirements.txt
torch 
torchvision 
torchaudio
--index-url https://download.pytorch.org/whl/cpu
numpy
Pillow
tqdm